commit return

+++ src/BmiService.java
@@ -3,6 +3,6 @@ public class BmiService {
     public double calculate(double weightKg, double heightM) {
         double index;
         index = weightKg / heightM / heightM;
-        return index;
+        return (int) index;
     }
 }
